Adding if condition for username

+++ assembl/graphql/user.py
@@ -235,7 +235,8 @@ class DeleteUserInformation(graphene.Mutation):
 
             # Deleting Username
             username = db.query(m.Username).filter(m.Username.user_id == user.id).first()
-            db.delete(username)
+            if username:
+                db.delete(username)
             # Delete Email Accounts
             email_account_ids = db.query(m.EmailAccount.id).join(m.User).filter(m.User.id == user.id).all()
             email_account_ids = [id for (id,) in email_account_ids]
